Initial Commit
So, I got a little lazy with this project and went ahead and did most of the ground work without making a repo for this game jam project.

I'm participating in the Major Jam 4: Cosmic game jam that ends at 11AM on August 16th 2021. This is the game that I'm working on.

This game is essentially a run and gun platformer, where you quickly input console commands to cast spells that temporarily buff your character or unleashes massive attacks. This is to fit with the theme of having to cheat to win the game.

So far, the ground work that I've done. By the time of this commit, this is day two of six.

Player character:
- Copied the 2D character controller script from brackeys and used that to help control the player.
-  I added the 'Better Jump' script, to make the player's jumping shorter and higher depending on the length of the button press. This code has been copied from that youtube video explaining this very topic.
- The player can shoot out simple bullets from its front, firing it in the direction it's facing.

- The health script has been reused from previous projects (Project Gun)

- The Game Event system has been reused from a previous project (let-go)

- Made editor tools for making objects that send the player to a specific place when entering its trigger collider.

- Made a simple OnTriggerEnterTrigger system, where other scripts can just subscribe to the event within the OnTriggerEnterTrigger script, and then that function will be invoked when an object enters the trigger collider.
- With this system, I have made a simple hurt the player when touched script.

- Enemies that patrol based on waypoints placed in the scene can be placed easily using editor tools.

The Command Line System:
The main meat of the game.

Today, I've finished the basics of it:
- The player can input the command by pressing right mouse button and then typing the command in the appeared input field.
- That input gets passed into the CommandExecutor script, where the input is compared against a list of Commands contained within the CommandList scriptable object. After confirming that the command is valid and exists within the list, the effect of the command will be run inside the coroutine of the CommandExecutor script.
- Other information regarding the execution is exported to the Scriptable object Variables, so that they may be displayed on the HUD.